# src/openapi.yaml
openapi: 3.1.0
info:
  title: MCP REST Bridge
  version: 1.0.0
  description: >
    A self-hosted MCP gateway that bridges MCP clients to backend REST APIs and
    other MCP servers, bidirectionally. REST-capable services register
    themselves (via OpenAPI/Swagger auto-discovery, a cURL/Postman import, or a
    manual tool list) and the bridge exposes their tools to any MCP client
    (e.g. Claude Desktop, Cursor, or a custom agent) via the Streamable HTTP
    MCP transport; existing MCP servers can likewise be registered as
    upstreams and re-exposed. Every call is proxied through a uniform guard
    pipeline (SSRF check, guardrails, per-tool policy, circuit breaker,
    response sanitizing, audit) documented in this admin API — this
    specification covers that management surface, not the dynamic per-client
    tool endpoints it provisions.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /admin-api/ws-proxy-targets:
    get:
      operationId: listWsProxyTargets
      summary: List WebSocket proxy targets
      tags: [WS Proxy]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      responses:
        "200":
          description: All targets, with live connection counts.
          content:
            {
              application/json:
                {
                  schema:
                    {
                      type: object,
                      properties: { items: { type: array, items: { $ref: "#/components/schemas/WsProxyTarget" } } },
                    },
                },
            }
    post:
      operationId: createWsProxyTarget
      summary: Create a new WebSocket proxy target
      description: The backend URL is SSRF-validated and IP-pinned; the target then becomes reachable at the `/ws-proxy/{name}` upgrade route (wired separately from Express, via the HTTP server's `upgrade` event). Requires the admin role for session callers.
      tags: [WS Proxy]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, backendWsUrl]
              properties:
                name: { type: string, pattern: "^[a-z0-9][a-z0-9_-]{0,62}$" }
                backendWsUrl: { type: string, example: "wss://backend.internal/socket" }
                maxConnections: { type: integer, minimum: 1 }
                maxMessageBytes: { type: integer, minimum: 1 }
                idleTimeoutMs: { type: integer, minimum: 1 }
                enabled: { type: boolean, default: true }
      responses:
        "201":
          description: Created.
          content: { application/json: { schema: { $ref: "#/components/schemas/WsProxyTarget" } } }
        "400":
          description: Invalid name, missing/invalid backendWsUrl, or an invalid numeric field.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "409":
          description: name collides with an existing WS proxy target (or, per the reciprocal check in register.ts, with an existing registered client).
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /admin-api/ws-proxy-targets/{name}:
    get:
      operationId: getWsProxyTarget
      summary: Get one WebSocket proxy target
      tags: [WS Proxy]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The target.
          content: { application/json: { schema: { $ref: "#/components/schemas/WsProxyTarget" } } }
        "404":
          description: No such target.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    patch:
      operationId: updateWsProxyTarget
      summary: Update a WebSocket proxy target
      description: Merged with the existing config — omitted fields keep their current value. Re-validates backendWsUrl (SSRF + IP pin) even when unchanged. Requires the admin role for session callers.
      tags: [WS Proxy]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                backendWsUrl: { type: string }
                maxConnections: { type: integer, minimum: 1 }
                maxMessageBytes: { type: integer, minimum: 1 }
                idleTimeoutMs: { type: integer, minimum: 1 }
                enabled: { type: boolean }
      responses:
        "200":
          description: Updated.
          content: { application/json: { schema: { $ref: "#/components/schemas/WsProxyTarget" } } }
        "400":
          description: Invalid backendWsUrl or numeric field.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "404":
          description: No such target.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    delete:
      operationId: deleteWsProxyTarget
      summary: Delete a WebSocket proxy target
      description: Requires the admin role for session callers.
      tags: [WS Proxy]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Deleted.
          content:
            {
              application/json:
                {
                  schema:
                    {
                      type: object,
                      properties: { status: { type: string, example: deleted }, name: { type: string } },
                    },
                },
            }
        "404":
          description: No such target.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
...
components:
...
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      description: >
        API key for administrative endpoints (registration, introspection).
        Set via ADMIN_API_KEYS env var. Disabled when AUTH_DISABLED=true.
    McpAuth:
      type: http
      scheme: bearer
      description: >
        API key for the backend-tool MCP transport endpoints
        (/mcp/{clientName}, /mcp-custom/{bundleName}). Set via MCP_API_KEYS
        env var, or a managed key (see MCP Keys). Disabled when
        AUTH_DISABLED=true or when no keys are configured at all.
    RootMcpAuth:
      type: http
      scheme: bearer
      description: >
        Credential for the /mcp system control-plane. Either the env admin
        Bearer key (ADMIN_API_KEYS) or a managed MCP key with `adminRole`
        set (see MCP Keys). Unlike McpAuth, there is no "no auth material
        configured => allow all" fallback — a caller must resolve to a real
        system role or the request is rejected outright, even on a fresh
        install with zero keys minted. Still bypassed entirely when
        AUTH_DISABLED=true (the same global dev/test escape hatch every
        other auth path in this API respects).
    SessionAuth:
      type: apiKey
      in: cookie
      name: __Host-mcp_admin_session
      description: >
        Admin-UI session cookie, issued by POST /admin-api/auth/login.
        Accepted anywhere AdminAuth is accepted. On mutating requests
        (anything but GET/HEAD/OPTIONS) an `X-CSRF-Token` header matching the
        `__Host-mcp_admin_csrf` cookie value is also required — OpenAPI 3.1
        has no first-class way to express that pairing, so it's documented
        here in prose, mirroring how RegistrationPayload's tools/openapi_url
        mutual-exclusivity is documented rather than schema-encoded.
        The `__Host-` prefix (and therefore this exact name) only applies
        when SESSION_COOKIE_SECURE=true (the default) — it mandates the
...
  schemas:
...
    WsProxyTarget:
      type: object
      properties:
        name: { type: string }
        backendWsUrl: { type: string }
        resolvedIp: { type: string, description: IP the backend hostname was pinned to at config time. }
        maxConnections: { type: integer }
        maxMessageBytes: { type: integer }
        idleTimeoutMs: { type: integer }
        enabled: { type: boolean }
        activeConnections: { type: integer }
        createdAt: { type: integer, description: Unix ms epoch }
        updatedAt: { type: integer, description: Unix ms epoch }

    # -------------------------------------------------------------------------
    # Catalog / marketplace
    # -------------------------------------------------------------------------
...
    ApiError:
      type: object
      description: >
        Standard error envelope returned by the bridge's REST endpoints (not
        MCP transport endpoints, which use JSON-RPC 2.0 error objects instead).
      required:
        - error
      properties:
        error:
          type: object
          description: Container for the structured error detail.
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: >
                Machine-readable error code in SCREAMING_SNAKE_CASE.
                Known values: VALIDATION_ERROR, CLIENT_NOT_FOUND, DISCOVERY_ERROR,
                UNAUTHORIZED, FORBIDDEN, RATE_LIMITED.
              example: VALIDATION_ERROR
            message:
              type: string
              description: >
                Human-readable description of what went wrong, suitable for
                displaying in a developer console or log.
              example: "Missing required fields: name, health_url"
            retry_after:
              type: integer
              description: >
                Seconds the caller should wait before retrying. Present only on
                a 429 RATE_LIMITED response (also echoed in the Retry-After
                header); omitted on every other error.
              example: 45
            request_id:
              type: [string, "null"]
              description: >
                Correlation id for this request (from res.locals.requestId),
                echoed back so a caller can quote it when reporting an issue.
                Present on almost every error response in the admin API;
                absent (or omitted) on the handful of handlers that don't
                thread it through (notably GET /clients and GET
                /clients/{name}/tools's error bodies, and /metrics's disabled
                response).
              example: "3f9c9a2e-9c34-4b3a-9a0e-4f7a6f6c9a1b"
      example:
        error:
          code: VALIDATION_ERROR
          message: "Missing required fields: name, health_url"
          request_id: "3f9c9a2e-9c34-4b3a-9a0e-4f7a6f6c9a1b"

    # -------------------------------------------------------------------------
    # JsonRpcMessage  (internal — MCP transport)
    # -------------------------------------------------------------------------
...